add kafka chart; add dependency mgmt; use find not ls
kafka chart deploys the kafka messaging platform as a petset.  also
deploys kafka-monitor for constant excercising of the platform

dependencies can be dynamically specified and packaged within helm
Kafka uses this functionality so we need to support it in the
makefile.

ls does not handle directories with spaces in their name correctly
find does